+++ changelog.md
@@ -2,6 +2,8 @@
 - Item editor now has a delete option to delete single items by right click
 - Added Skill and Luck to the item creation form
 - Made some UI improvments to the warehouse view to better match MU
+- Added 5th wing options to the item creator
+- Added Harmony Options to the item creator
 
 # [v0.0.2]
 - Added warehouse inventory decoder for Base64-encoded item data